Move Node from DFSSolver to the model package
